Virtualbox: move NAT port allocation to machine boot
The host side ssh NAT port was being chosen by the OS
when the VM was crafted in the `Create` method.
There was no accounting for some other process taking
over the port while the VM is shutoff.

This moves the port provisioning to VM start.
On first start, the OS will supply the port.  Subsequent
starts, we see if we can re-use the port, falling back
to delegating the selection to the OS.

Signed-off-by: John Fieber <jrf@ursamaris.org>