Add VARIANT and VARIANT_ID to os-release
In order to provision different variant types such as embedded and
server releases, it's required to grep the VARIANT and VARIANT_ID
variables of os-release.

Specifically from this commit:
https://pagure.io/fork/sgallagh/fedora-release/c/56267336c738f80e3e7a344c740b5d02730c76c6

Which adds the VARIANT_ID="atomic.host" in order to differentiate
between Fedora Server and container-centric Atomic Host releases.

This commit adds the accessability of VARIANT and VARIANT_ID in order
for Docker Machine to successfully provision different variants of Fedora
releases.

Signed-off-by: Charlie Drage <charlie@charliedrage.com>