Remove hard-coded incorrect "fish" shell identification.
The lines removed cause the shell to be identified as `fish` if the `__fish_bin_dir`, even when `$SHELL` is explicitly set to something else.

This causes a fish user to get the `fish` env when running a non-fish sub-shell.  All bash scripts that use `docker-machine env` will fail for fish users.

Signed-off-by: Scott M. Smith <scottsmith1@gmail.com>